guess game but keep a count
An array named record_attempts is added to store the number of attempts made by the user. The size of the array is set to 1000, which should be sufficient for most games.
When the user correctly guesses the number, the program records the number of attempts in the record_attempts array.
After the game is finished, the program displays a record of all the attempts made by the user, showing the attempt number and the corresponding number of attempts it took to guess the correct number.